+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/StoreResponse.java
@@ -11,9 +11,9 @@ import com.azure.cosmos.implementation.directconnectivity.rntbd.RntbdEndpointSta
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
-import com.azure.cosmos.implementation.Utils;
-import java.util.concurrent.ConcurrentSkipListMap;
 
 /**
  * Used internally to represents a response from the store.
@@ -21,7 +21,8 @@ import java.util.concurrent.ConcurrentSkipListMap;
 public class StoreResponse {
     final static Logger LOGGER = LoggerFactory.getLogger(StoreResponse.class);
     final private int status;
-    final private Map<String, String> responseHeaders;
+    final private String[] responseHeaderNames;
+    final private String[] responseHeaderValues;
     final private byte[] content;
 
     private int pendingRequestQueueSize;
@@ -33,6 +34,7 @@ public class StoreResponse {
     private RntbdEndpointStatistics rntbdEndpointStatistics;
     private int rntbdRequestLength;
     private int rntbdResponseLength;
+    private final List<String> replicaStatusList;
 
     public StoreResponse(
             int status,
@@ -40,22 +42,35 @@ public class StoreResponse {
             byte[] content) {
 
         requestTimeline = RequestTimeline.empty();
-        responseHeaders = new ConcurrentSkipListMap<>(String.CASE_INSENSITIVE_ORDER);
-        responseHeaders.putAll(headerMap);
+        responseHeaderNames = new String[headerMap.size()];
+        responseHeaderValues = new String[headerMap.size()];
+
+        int i = 0;
+        for (Map.Entry<String, String> headerEntry : headerMap.entrySet()) {
+            responseHeaderNames[i] = headerEntry.getKey();
+            responseHeaderValues[i] = headerEntry.getValue();
+            i++;
+        }
 
         this.status = status;
         this.content = content;
         if (this.content != null) {
             this.responsePayloadLength = this.content.length;
         }
+
+        replicaStatusList = new ArrayList<>();
     }
 
     public int getStatus() {
         return status;
     }
 
-    public Map<String, String> getResponseHeaders() {
-        return responseHeaders;
+    public String[] getResponseHeaderNames() {
+        return responseHeaderNames;
+    }
+
+    public String[] getResponseHeaderValues() {
+        return responseHeaderValues;
     }
 
     public int getRntbdChannelTaskQueueSize() {
@@ -128,7 +143,17 @@ public class StoreResponse {
     }
 
     public String getHeaderValue(String attribute) {
-        return responseHeaders.get(attribute);
+        if (this.responseHeaderValues == null || this.responseHeaderNames.length != this.responseHeaderValues.length) {
+            return null;
+        }
+
+        for (int i = 0; i < responseHeaderNames.length; i++) {
+            if (responseHeaderNames[i].equalsIgnoreCase(attribute)) {
+                return responseHeaderValues[i];
+            }
+        }
+
+        return null;
     }
 
     public double getRequestCharge() {
@@ -178,5 +203,9 @@ public class StoreResponse {
             }
         }
         return subStatusCode;
+    }
+
+    public List<String> getReplicaStatusList() {
+        return this.replicaStatusList;
     }
 }