+++ sdk/formrecognizer/azure-ai-formrecognizer/src/main/java/com/azure/ai/formrecognizer/documentanalysis/models/DocumentTable.java
@@ -4,12 +4,14 @@
 package com.azure.ai.formrecognizer.documentanalysis.models;
 
 import com.azure.ai.formrecognizer.documentanalysis.implementation.util.DocumentTableHelper;
+import com.azure.core.annotation.Immutable;
 
 import java.util.List;
 
 /**
  * A table object consisting table cells arranged in a rectangular layout.
  */
+@Immutable
 public final class DocumentTable {
     /*
      * Number of rows in the table.
@@ -51,7 +53,7 @@ public final class DocumentTable {
      * @param rowCount the rowCount value to set.
      * @return the DocumentTable object itself.
      */
-    void setRowCount(int rowCount) {
+    private void setRowCount(int rowCount) {
         this.rowCount = rowCount;
     }
 
@@ -70,7 +72,7 @@ public final class DocumentTable {
      * @param columnCount the columnCount value to set.
      * @return the DocumentTable object itself.
      */
-    void setColumnCount(int columnCount) {
+    private void setColumnCount(int columnCount) {
         this.columnCount = columnCount;
     }
 
@@ -89,7 +91,7 @@ public final class DocumentTable {
      * @param cells the cells value to set.
      * @return the DocumentTable object itself.
      */
-    void setCells(List<DocumentTableCell> cells) {
+    private void setCells(List<DocumentTableCell> cells) {
         this.cells = cells;
     }
 
@@ -108,7 +110,7 @@ public final class DocumentTable {
      * @param boundingRegions the boundingRegions value to set.
      * @return the DocumentTable object itself.
      */
-    void setBoundingRegions(List<BoundingRegion> boundingRegions) {
+    private void setBoundingRegions(List<BoundingRegion> boundingRegions) {
         this.boundingRegions = boundingRegions;
     }
 
@@ -127,7 +129,7 @@ public final class DocumentTable {
      * @param spans the spans value to set.
      * @return the DocumentTable object itself.
      */
-    void setSpans(List<DocumentSpan> spans) {
+    private void setSpans(List<DocumentSpan> spans) {
         this.spans = spans;
     }